backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1

# Google GenAI SDK (Gemini Live API)
google-genai==1.5.0

# Audio processing
numpy==2.2.1

# Image processing
Pillow==11.1.0

# Cloud services (for later days)
# google-cloud-firestore==2.19.0
# google-cloud-storage==2.19.0
# google-cloud-secret-manager==2.21.1

# CORS
python-multipart==0.0.18

# Logging
structlog==24.4.0
